requirements.txt
asgiref==3.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
Django==5.2.4
django-environ==0.12.0
django-mapbox-location-field==2.1.0
fontawesome-free==5.15.4
idna==3.10
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
sweetify==2.3.1
tzdata==2025.2
urllib3==2.5.0
